mqtt/requirements.txt
requests
paho-mqtt==1.6.0